Created S3Service for amazon s3
