+++ src/Traces.Core/Services/Files/ITraceFileService.cs
@@ -1,9 +1,6 @@
-using System;
 using System.Collections.Generic;
-using System.Linq.Expressions;
 using System.Threading.Tasks;
 using Traces.Core.Models.Files;
-using Traces.Data.Entities;
 
 namespace Traces.Core.Services.Files
 {
@@ -13,6 +10,6 @@ namespace Traces.Core.Services.Files
 
         Task<SavedFileDto> GetSavedFileFromPublicIdAsync(string publicId);
 
-        Task<bool> DeleteTraceFileRangeAsync(Expression<Func<TraceFile, bool>> expression);
+        Task<bool> DeleteTraceFileRangeAsync(List<int> ids);
     }
 }
