+++ src/Traces.Core/Services/Traces/TraceService.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using NodaTime;
 using Optional;
@@ -15,6 +16,7 @@ using Traces.Core.ClientFactories;
 using Traces.Core.Extensions;
 using Traces.Core.Models;
 using Traces.Core.Repositories;
+using Traces.Core.Services.Files;
 using Traces.Data.Entities;
 
 namespace Traces.Core.Services.Traces
@@ -24,9 +26,11 @@ namespace Traces.Core.Services.Traces
         private readonly ITraceRepository _traceRepository;
         private readonly IRequestContext _requestContext;
         private readonly IApaleoClientsFactory _apaleoClientsFactory;
+        private readonly ITraceFileService _traceFileService;
 
-        public TraceService(ITraceRepository traceRepository, IRequestContext requestContext, IApaleoClientsFactory apaleoClientsFactory)
+        public TraceService(ITraceRepository traceRepository, IRequestContext requestContext, IApaleoClientsFactory apaleoClientsFactory, ITraceFileService traceFileService)
         {
+            _traceFileService = Check.NotNull(traceFileService, nameof(traceFileService));
             _traceRepository = Check.NotNull(traceRepository, nameof(traceRepository));
             _requestContext = Check.NotNull(requestContext, nameof(requestContext));
             _apaleoClientsFactory = Check.NotNull(apaleoClientsFactory, nameof(apaleoClientsFactory));
@@ -239,6 +243,9 @@ namespace Traces.Core.Services.Traces
             {
                 throw new BusinessValidationException(string.Format(TextConstants.TraceCouldNotBeFoundErrorMessageFormat, id));
             }
+
+            var trace = await _traceRepository.GetAsync(id);
+            await _traceFileService.DeleteTraceFileRangeAsync(trace.Files.Select(tf => tf.Id).ToList());
 
             var deleted = await _traceRepository.DeleteAsync(id);
